Use dataclasses with custom managed properties
Use dataclasses instead of NamedTuples. These provide better facilities
for converting to json (dataclasses.asdict is recursive!), are mutable,
and support data validation (which is coming soon) via custom properties.

Also create a @managed_property decorator for use with dataclasses.
This decorator creates a property that is backed by a private property,
with an auto-generated getter and custom setter.

Notably, the managed property uses the value _returned by_ the setter,
instead of requiring the setter to directly update the private property.
This allows us to specify the name of the private property only once.