Allow parse_pargs() to pick up sys.argv
Make the `argv` parameter of main() optional. When the script is called
directly, don't pass any args to main() (instead of slicing sys.argv
manually). This allows parse_args() to fall back to using sys.argv.

This change was introduced so that I can use "animdata:main" as the
console_scripts entrypoint in setup.py.