👷Publish to Test PyPI when a release is created
1. Make the Build workflow also run when a new release is published
   through GitHub's web UI.
2. Create a separate publish step that runs only when the workflow is
   triggered by a release.

This allows me to control when to publish to PyPI.

Note: Previously, I blindly assumed that tox would build both sdists and
wheels before testing, and that I could publish these sdists and wheels
after testing. However:

1. tox builds sdists only, but not wheels. This is not configurable.
2. tox builds sdists using a ZIP format. A non-issue compared to 1.

Because of this, I am forced to manually build sdists and wheels during
the publish step.