Move out-of-bounds trigger check to CLI
The Record class no longer checks if a trigger frame is outside the
bounds defined by frames_per_direction. This logic seems too complicated
to put in a validator. Also, because it depends both on Record.triggers
and Record.frames_per_direction, it may interfere with construction of a
valid Record instance.

Instead, move this check into the CLI script. Create the
check_out_of_bounds_triggers() function, which is used by main() to
check records extracted from an AnimData.D2, tabbed text, or JSON. If a
record contains an out-of-bounds trigger, display a warning.